requirements.txt
fastapi
uvicorn
python-dotenv[standard]
boto3